Got rid of old interface, renamed DRListManager to DebrisCollection
