+++ src/main/java/debrisProcessingSubsystem/debrisCollection/DebrisCollection.java
@@ -1,44 +1,110 @@
 package debrisProcessingSubsystem.debrisCollection;
 
+import debrisProcessingSubsystem.Updatable;
+import debrisProcessingSubsystem.Update;
+
 /**
- * @deprecated Debris Collections will now implement the Updatable interface.
- * This will remain in place for testing.
- * This will be the interface implemented by the Debris Collection object shown
- * in the SADD.
- * This interface should contain all outward facing methods for the
- * Debris collection data structure.
- *
- * This is a preliminary placeholder and very subject to change.
- * Created by jdt on 3/4/17.
+ * Created by jdt on 3/13/17.
  */
-public interface DebrisCollection {
+public class DebrisCollection implements Updatable {
+
+    private DebrisList newDebris, oldDebris;
 
 
     /**
-     * This method will be called when it is time to add a new image to the
-     * debris list.
-     * @return True on success.
+     * Default constructor. Initializes debris lists to null.
      */
-    public boolean newImage();
+    public DebrisCollection(){
+        newDebris = null;
+        oldDebris = null;
+    }
 
     /**
-     * Add a new debris item.
-     * TODO this should take some debris parameters.
+     * Call this to send an update to this component.
+     * Returns a response Update
+     * TODO write this.
+     * @param theUpdate The update to be received by this component.
+     * @return A response update.
      */
-    public void addDebris();
+    public Update updateComponent(Update theUpdate){
+        return null;
+    }
 
     /**
-     * This method is to be called when. SADD team appears to envision this
-     * implemented with an iterator.
-     * possibly within the debris list itself?
+     * Return any available updates.
+     * TODO write this.
+     * @return An update is ready.
      */
-    public DebrisRecord getDebris();
+    public Update pollComponent(){
+        return null;
+    }
 
     /**
-     * Print some information about the collection.
+     * Start a new image.
+     * @return success or failure
+     */
+    private boolean newImage(){
+        //swap lists
+        if(newDebris == null){
+            newDebris = new DRList();
+        }
+        else {
+            //clear new lists
+            swapLists();
+            //ok to add images?
+            //track debris
+        }
+        return true;
+    }
+
+    /**
+     * Add a debris record to the list.
+     * TODO needs parameters.
+     */
+    private void addDebris(){
+        try {
+            newDebris.addDebris(new DebrisRecord());
+        }
+        catch(NullPointerException e){
+            System.err.println("New Debris list not initialized: " + e.getMessage());
+        }
+    }
+
+    /**
+     * Get a DebrisRecord from the old list for sending.
+     * @return Next DebrisRecord in the old debris list.
+     */
+    public DebrisRecord getDebris(){
+        return oldDebris.getDebrisElement();
+    }
+
+    /**
+     * Testing method
+     * Print some info about this collection.
      * Testing method.
      */
-    public void printCollectionCharacteristics();
+    public void printCollectionCharacteristics(){
+        if(newDebris != null){
+            System.out.println("newDebris: " + newDebris.size());
+        }
+        if(oldDebris != null){
+            System.out.println("oldDebris: " + oldDebris.size());
+        }
+    }
 
-    //methods swapList and clearList should be internal per prof roman.
+    /**
+     * Make newDebris the old list and create a new list for the next image.
+     */
+    private void swapLists(){
+        oldDebris = newDebris;
+        newDebris = new DRList();
+    }
+
+    /**
+     * Testing main for DebrisCollection.
+     * @param args ignored.
+     */
+    public static void main(String[] args){
+
+    }
 }
